Moved all .h files to include/optmizer
