+++ src/include/optimizer/column_manager.h
@@ -4,7 +4,7 @@
 //
 // column_manager.h
 //
-// Identification: src/backend/optimizer/column_manager.h
+// Identification: src/optimizer/column_manager.h
 //
 // Copyright (c) 2015-16, Carnegie Mellon University Database Group
 //
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include "backend/optimizer/column.h"
+#include "optimizer/column.h"
 
 namespace peloton {
 namespace optimizer {
