+++ src/include/optimizer/optimizer.h
@@ -4,7 +4,7 @@
 //
 // optimizer.h
 //
-// Identification: src/backend/optimizer/optimizer.h
+// Identification: src/optimizer/optimizer.h
 //
 // Copyright (c) 2015-16, Carnegie Mellon University Database Group
 //
@@ -12,17 +12,17 @@
 
 #pragma once
 
-#include "backend/optimizer/memo.h"
-#include "backend/optimizer/column_manager.h"
-#include "backend/optimizer/query_operators.h"
-#include "backend/optimizer/operator_node.h"
-#include "backend/optimizer/binding.h"
-#include "backend/optimizer/pattern.h"
-#include "backend/optimizer/property.h"
-#include "backend/optimizer/property_set.h"
-#include "backend/optimizer/rule.h"
-#include "backend/planner/abstract_plan.h"
-#include "backend/common/logger.h"
+#include "optimizer/memo.h"
+#include "optimizer/column_manager.h"
+#include "optimizer/query_operators.h"
+#include "optimizer/operator_node.h"
+#include "optimizer/binding.h"
+#include "optimizer/pattern.h"
+#include "optimizer/property.h"
+#include "optimizer/property_set.h"
+#include "optimizer/rule.h"
+#include "planner/abstract_plan.h"
+#include "common/logger.h"
 
 #include <memory>
