+++ src/include/optimizer/query_node_visitor.h
@@ -4,7 +4,7 @@
 //
 // query_node_visitor.h
 //
-// Identification: src/backend/optimizer/query_node_visitor.h
+// Identification: src/optimizer/query_node_visitor.h
 //
 // Copyright (c) 2015-16, Carnegie Mellon University Database Group
 //
