+++ src/include/optimizer/stats.h
@@ -2,9 +2,9 @@
 //
 //                         Peloton
 //
-// op_expression.h
+// stats.h
 //
-// Identification: src/backend/optimizer/op_expression.h
+// Identification: src/optimizer/stats.h
 //
 // Copyright (c) 2015-16, Carnegie Mellon University Database Group
 //
@@ -12,35 +12,20 @@
 
 #pragma once
 
-#include "backend/optimizer/operator_node.h"
-#include "backend/optimizer/group.h"
-
-#include <vector>
-#include <memory>
+#include "optimizer/tuple_sample.h"
 
 namespace peloton {
 namespace optimizer {
 
 //===--------------------------------------------------------------------===//
-// Operator Expr
+// Stats
 //===--------------------------------------------------------------------===//
-class OpExpressionVisitor;
+class Stats {
+ public:
+  Stats(TupleSample *sample);
 
-class OpExpression {
-public:
-  OpExpression(Operator op);
-
-  void PushChild(std::shared_ptr<OpExpression> op);
-
-  void PopChild();
-
-  const std::vector<std::shared_ptr<OpExpression>> &Children() const;
-
-  const Operator &Op() const;
-
-private:
-  Operator op;
-  std::vector<std::shared_ptr<OpExpression>> children;
+ private:
+  TupleSample *sample;
 };
 
 } /* namespace optimizer */
