+++ src/optimizer/convert_query_to_op.cpp
@@ -4,23 +4,23 @@
 //
 // convert_query_to_op.cpp
 //
-// Identification: src/backend/optimizer/convert_query_to_op.cpp
+// Identification: src/optimizer/convert_query_to_op.cpp
 //
 // Copyright (c) 2015-16, Carnegie Mellon University Database Group
 //
 //===----------------------------------------------------------------------===//
 
-#include "backend/optimizer/convert_query_to_op.h"
-#include "backend/optimizer/query_node_visitor.h"
-#include "backend/optimizer/operators.h"
+#include "optimizer/convert_query_to_op.h"
+#include "optimizer/query_node_visitor.h"
+#include "optimizer/operators.h"
 
-#include "backend/planner/projection_plan.h"
-#include "backend/planner/seq_scan_plan.h"
-#include "backend/planner/order_by_plan.h"
+#include "planner/projection_plan.h"
+#include "planner/seq_scan_plan.h"
+#include "planner/order_by_plan.h"
 
-#include "backend/catalog/manager.h"
-#include "backend/bridge/ddl/bridge.h"
-#include "backend/bridge/dml/mapper/mapper.h"
+#include "catalog/manager.h"
+#include "bridge/ddl/bridge.h"
+#include "bridge/dml/mapper/mapper.h"
 
 namespace peloton {
 namespace optimizer {
