+++ src/optimizer/operator_visitor.cpp
@@ -4,13 +4,13 @@
 //
 // operator_visitor.cpp
 //
-// Identification: src/backend/optimizer/operator_visitor.cpp
+// Identification: src/optimizer/operator_visitor.cpp
 //
 // Copyright (c) 2015-16, Carnegie Mellon University Database Group
 //
 //===----------------------------------------------------------------------===//
 
-#include "backend/optimizer/operator_visitor.h"
+#include "optimizer/operator_visitor.h"
 
 namespace peloton {
 namespace optimizer {
