+++ src/optimizer/pattern.cpp
@@ -4,13 +4,13 @@
 //
 // pattern.cpp
 //
-// Identification: src/backend/optimizer/pattern.cpp
+// Identification: src/optimizer/pattern.cpp
 //
 // Copyright (c) 2015-16, Carnegie Mellon University Database Group
 //
 //===----------------------------------------------------------------------===//
 
-#include "backend/optimizer/pattern.h"
+#include "optimizer/pattern.h"
 
 namespace peloton {
 namespace optimizer {
