+++ src/optimizer/property_set.cpp
@@ -4,13 +4,13 @@
 //
 // property_set.cpp
 //
-// Identification: src/backend/optimizer/property_set.cpp
+// Identification: src/optimizer/property_set.cpp
 //
 // Copyright (c) 2015-16, Carnegie Mellon University Database Group
 //
 //===----------------------------------------------------------------------===//
 
-#include "backend/optimizer/property_set.h"
+#include "optimizer/property_set.h"
 
 namespace peloton {
 namespace optimizer {
