+++ src/optimizer/stats.cpp
@@ -4,13 +4,13 @@
 //
 // stats.cpp
 //
-// Identification: src/backend/optimizer/stats.cpp
+// Identification: src/optimizer/stats.cpp
 //
 // Copyright (c) 2015-16, Carnegie Mellon University Database Group
 //
 //===----------------------------------------------------------------------===//
 
-#include "backend/optimizer/stats.h"
+#include "optimizer/stats.h"
 
 namespace peloton {
 namespace optimizer {
